Parking lot project document overview details and written information
This document was all the information i managed to put together in the document. This includes everything about what i managed to see in the data of the USB drive. I also, add more details in terms of what could of been malicious of its nature and give more details on what the files that were found that could present vulnerability risks of the situation.